refactor(midway-mock): fix types path (#337)
