+++ packages/midway-mock/bootstrap.d.ts
@@ -1,10 +1,2 @@
 
-import * as assert from 'power-assert';
-import { MidwayMock, MidwayMockApplication } from './src';
-
-export {
-  assert
-};
-export declare const app: MidwayMockApplication;
-export declare const mock: MidwayMock;
-export declare const mm: MidwayMock;
+export * from './dist/bootstrap'
